+++ scripts/install-manjaro.sh
@@ -138,6 +138,8 @@ manjaro-chroot /mnt ${CMD_PACMAN_INSTALL} lutris
 manjaro-chroot /mnt sudo -u stick yay --noconfirm -S heroic-games-launcher-bin
 # Steam.
 manjaro-chroot /mnt ${CMD_PACMAN_INSTALL} gcc-libs libgpg-error libva libxcb lib32-gcc-libs lib32-libgpg-error lib32-libva lib32-libxcb steam-manjaro steam-native
+# protontricks.
+manjaro-chroot /mnt sudo -u stick yay --noconfirm -S protontricks
 # Proton GE for Steam.
 wget https://raw.githubusercontent.com/toazd/ge-install-manager/master/ge-install-manager -O /mnt/usr/local/bin/ge-install-manager
 chmod +x /mnt/usr/local/bin/ge-install-manager