+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add Linux 5.4 as an alternative LTS kernel to 5.10.
 - Change GRUB menu to show on boot.
 - Add Heroic Games Launcher for Epic Games Store games.
+- Change partition mounts from UUIDs to labels.
 
 ## 2.0.0