+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Add 16 GB exFAT partition for normal flash drive usage.
 - Add PulseAudio configuration to lower the volume which will helps Mac users.
 - Add printer drivers and service.
+- Add Oh My Zsh.
 
 ## 2.0.0